+++ back-end/logs/user_interactions.csv
@@ -16,3 +16,4 @@ timestamp,interaction_id,user_query,extracted_origin,extracted_destination,sugge
 ,658e94b5-47db-4592-afc8-505439f1970a,meu ônibus atrasou,,,,,"The best route is the Linha 51, which goes from DIADEMA (VILA SAO JOSE) to SAUDE in São Paulo.",
 ,7377126f-0f1a-4111-9af4-6eea649d5418,céu de Hortolândia quero ir para,,,,,"The best route is the Linha 746, which goes from Hortolândia to Campinas.",feedback_positivo_concluido: obrigado
 ,8fa5f559-e36c-4f1d-a729-11cb57ec08c8,obrigado obrigado,,,,,"“A melhor rota é a Linha 487, que vai de Monte Mor até Santo André.”",
+,46b8d221-1c61-4cb4-ba4d-109a26a65f65,lalala,,,,,"The best route is the Linha 487, which goes from MAUÁ to Santo André.",
