+++ back-end/main.py
@@ -2,9 +2,8 @@ import os
 import pandas as pd
 from dotenv import load_dotenv
 from config import constants
-from core.routes import RouteFetcher
-from model.gemini_analysis import GeminiThinking
-from model.llama_analysis import LlamaThinking
+from core import RouteFetcher
+from model import GeminiThinking, LlamaThinking, BusSenseAssistant
 import time
 
 
@@ -23,7 +22,7 @@ if __name__ == "__main__":
         #     "Enter your choice (1 or 2): "
         # )
 
-        assistant_choose = "2"
+        assistant_choose = "3"
 
         if assistant_choose == "1":
             print('gemini')
@@ -31,6 +30,10 @@ if __name__ == "__main__":
         elif assistant_choose == "2":
             assistant = LlamaThinking(routes_data=routes_df)
             print('llama')
+        elif assistant_choose == "3":
+            assistant = BusSenseAssistant(routes_data_path=constants.DATA_FILE_PATH + 'llm_generated_routes.csv')
+            assistant.start()
+            break
         else:
             print("\nEscolha apenas 1 ou 2.\n")
             time.sleep(1)
